// config/client/models/ContactReadDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ContactReadDto = {
    fullname: string;
    email?: (string | null);
    phone: string;
    title: string;
    message?: (string | null);
    id: number;
};